train_probe.py
"""
Train a probe on RepE contrastive activations and validate on MASK.

Usage:
    python train_probe.py --config configs/cadenza.yaml

    # Override specific options
    python train_probe.py --config configs/cadenza.yaml --method mms

Validation metric: TPR on MASK at the threshold giving 1% FPR on alpaca.
"""

import argparse
from pathlib import Path

import numpy as np
import yaml
from sklearn.metrics import roc_auc_score

from probes import get_probe_class
from probes.base import BaseProbe


def compute_tpr_at_fpr(
    scores_positive: np.ndarray,
    scores_negative: np.ndarray,
    target_fpr: float = 0.01,
) -> tuple[float, float]:
    """Compute TPR at a given FPR threshold.

    Args:
        scores_positive: Scores for deceptive samples (should be high).
        scores_negative: Scores for honest/benign samples (should be low).
        target_fpr: Target false positive rate.

    Returns:
        (tpr, threshold): TPR achieved at the threshold giving the target FPR.
    """
    # Threshold is set such that target_fpr fraction of negatives exceed it
    threshold = np.percentile(scores_negative, 100 * (1 - target_fpr))
    tpr = (scores_positive >= threshold).mean()
    return float(tpr), float(threshold)


def load_activations(
    output_dir: Path, dataset_name: str, model_name: str, layer: int,
) -> tuple[np.ndarray, np.ndarray]:
    """Load cached activations. Raises FileNotFoundError if not cached."""
    safe_model = model_name.replace("/", "_")
    path = output_dir / f"{dataset_name}_{safe_model}_layer{layer}.npz"
    if not path.exists():
        raise FileNotFoundError(
            f"Activations not found: {path}\n"
            f"Run: python extract_activations.py --config <config> --dataset {dataset_name}"
        )
    data = np.load(path)
    return data["activations"], data["labels"]


def train(
    method: str,
    model_name: str,
    layer: int,
    activations_dir: Path,
    output_dir: Path,
    val_fraction: float = 0.2,
    val_alpaca_size: int = 200,
    **probe_kwargs,
) -> BaseProbe:
    """Train a probe and validate.

    Returns the trained probe.
    """
    # --- Load RepE training data ---
    acts, labels = load_activations(activations_dir, "repe", model_name, layer)
    if len(acts) == 0:
        raise RuntimeError("RepE activations are empty. Did you extract them?")

    pos_acts = acts[labels]       # deceptive
    neg_acts = acts[~labels]      # honest
    print(f"RepE data: {len(pos_acts)} deceptive, {len(neg_acts)} honest")

    # Split into train / held-out val
    n_val = int(len(pos_acts) * val_fraction)
    if n_val > 0:
        train_pos, val_pos = pos_acts[:-n_val], pos_acts[-n_val:]
        train_neg, val_neg = neg_acts[:-n_val], neg_acts[-n_val:]
    else:
        train_pos, val_pos = pos_acts, np.array([])
        train_neg, val_neg = neg_acts, np.array([])

    # --- Train ---
    probe_cls = get_probe_class(method)
    probe = probe_cls(**probe_kwargs)
    print(f"Training {probe_cls.__name__} (method={method})...")
    probe.fit(train_pos, train_neg)

    # --- Validate on held-out RepE ---
    if len(val_pos) > 0:
        val_scores = probe.score(np.vstack([val_pos, val_neg]))
        val_labels = np.concatenate([np.ones(len(val_pos)), np.zeros(len(val_neg))])
        val_auroc = roc_auc_score(val_labels, val_scores)
        print(f"RepE held-out AUROC: {val_auroc:.4f}")

    # --- Validate on MASK (TPR @ 1% FPR on alpaca) ---
    try:
        mask_acts, mask_labels = load_activations(activations_dir, "mask", model_name, layer)
        alpaca_acts, alpaca_labels = load_activations(activations_dir, "alpaca", model_name, layer)
    except FileNotFoundError as e:
        print(f"Skipping MASK validation: {e}")
        mask_acts, alpaca_acts = np.array([]), np.array([])

    if len(mask_acts) > 0 and len(alpaca_acts) > 0:
        # Subsample alpaca for validation (default 200, separate from the full
        # set used during evaluation which defaults to 2000)
        if val_alpaca_size and len(alpaca_acts) > val_alpaca_size:
            rng = np.random.RandomState(42)
            val_alpaca_idx = rng.choice(len(alpaca_acts), val_alpaca_size, replace=False)
            alpaca_acts_val = alpaca_acts[val_alpaca_idx]
        else:
            alpaca_acts_val = alpaca_acts

        mask_scores = probe.score(mask_acts)
        alpaca_scores = probe.score(alpaca_acts_val)

        # TPR (accuracy on all-deceptive MASK) at 1% FPR calibrated on alpaca
        tpr, threshold = compute_tpr_at_fpr(mask_scores, alpaca_scores, target_fpr=0.01)
        # Accuracy at that threshold: fraction of MASK correctly flagged as deceptive
        accuracy_at_threshold = float((mask_scores >= threshold).mean())

        # AUROC treating MASK as positive, alpaca as negative
        combined_scores = np.concatenate([mask_scores, alpaca_scores])
        combined_labels = np.concatenate([np.ones(len(mask_scores)), np.zeros(len(alpaca_scores))])
        mask_auroc = roc_auc_score(combined_labels, combined_scores)

        print(f"MASK validation ({len(mask_scores)} MASK, {len(alpaca_scores)} Alpaca):")
        print(f"  AUROC (MASK vs Alpaca): {mask_auroc:.4f}")
        print(f"  TPR @ 1% FPR:          {tpr:.4f} (threshold={threshold:.4f})")
        print(f"  Accuracy @ 1% FPR:     {accuracy_at_threshold:.4f}")

    # --- Save ---
    output_dir.mkdir(parents=True, exist_ok=True)
    probe_path = output_dir / f"probe_{method}_{model_name}_layer{layer}.pkl"
    probe.save(probe_path)
    print(f"Saved probe to: {probe_path}")

    return probe


def main():
    parser = argparse.ArgumentParser(description="Train a deception probe")
    parser.add_argument("--config", type=str, required=True, help="Path to YAML config")
    parser.add_argument("--method", type=str, help="Override probe method from config")
    parser.add_argument("--activations_dir", type=str, help="Override activations directory")
    parser.add_argument("--output_dir", type=str, help="Override output directory")
    args = parser.parse_args()

    with open(args.config) as f:
        cfg = yaml.safe_load(f)

    method = args.method or cfg["method"]
    model_name = cfg["model_name"]
    layer = cfg["detect_layers"][0]
    activations_dir = Path(args.activations_dir or cfg.get("activations_dir", "activations"))
    output_dir = Path(args.output_dir or cfg.get("output_dir", "outputs"))
    val_fraction = cfg.get("val_fraction", 0.2)
    val_alpaca_size = cfg.get("val_alpaca_size", 200)

    # Collect probe-specific kwargs from config
    probe_kwargs = {}
    if method == "lr":
        probe_kwargs["reg_coeff"] = cfg.get("reg_coeff", 10.0)
        probe_kwargs["max_iter"] = cfg.get("max_iter", 1000)

    train(
        method=method,
        model_name=model_name,
        layer=layer,
        activations_dir=activations_dir,
        output_dir=output_dir,
        val_fraction=val_fraction,
        val_alpaca_size=val_alpaca_size,
        **probe_kwargs,
    )


if __name__ == "__main__":
    main()
