pyproject.toml
[project]
name = "probing"
version = "0.1.0"
description = "Lightweight deception probing mini-repo for liars-bench"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "datasets>=2.19",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "accelerate>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
